+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Model derivative downloads now use the new direct-s3 logic.
+
 ## [9.0.0] - 2022-08-19
 
-# Changed
+### Changed
 - Uploading/downloading of OSS objects now uses the new binary transfer endpoints.
 - Resumable upload methods (`uploadObjectResumable`, `uploadObjectStreamResumable`, `getResumableUploadStatus`)
 are now marked as deprecated, and will be removed in the next major release. Use `getUploadUrls` and `completeUpload` instead.